pyproject.toml
[project]
# --- Core Project Metadata ---
name = "cubesat-virtual-testbed"
version = "0.1.0"
description = "A high-fidelity digital twin for CubeSat simulation, focusing on physics and on-board AI constraints."
readme = "README.md"
requires-python = ">=3.10" # We are targeting 3.11

# --- Primary Dependencies ---
dependencies = [
    "fastapi>=0.96,<1.0",
    "uvicorn[standard]>=0.25,<1.0",

    "pydantic>=2.4,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",

    # --- Physics, Math & Orbit (Pinned for Python 3.11) ---
    "numpy~=1.26",      
    "scipy~=1.13",      
    "astropy>=5.0,<6.0",
    "poliastro~=0.17",
    "tle-tools>=0.2.3"  
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "httpx>=0.26,<0.30",
    "black>=24.0,<25.0",
]

ai_tflite = [
    "tensorflow-lite-runtime>=2.13,<3.0",
]

ai_onnx = [
    "onnxruntime>=1.15,<2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["cvt*"]
